% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_authors.R
\name{apply_grepl}
\alias{apply_grepl}
\title{apply_grepl}
\usage{
apply_grepl(
  x,
  ...,
  other = FALSE,
  ignore.case = TRUE,
  perl = FALSE,
  fixed = FALSE
)
}
\arguments{
\item{x}{vector of strings to modify}

\item{...}{list syntax, replacement = c("words", "to", "replace")}

\item{other}{if TRUE, items not replaced by user commands will be replaced by "Other"}

\item{ignore.case}{if TRUE, case is not matched}
}
\description{
Batch together grepl commands
}
\details{
using NULL as a replacement string will result in NAs
}
