% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreview_oracle.R
\name{asreview.csv_to_ris}
\alias{asreview.csv_to_ris}
\title{oracle.csv_to_ris}
\usage{
asreview.csv_to_ris(csv, file)
}
\arguments{
\item{csv}{path to the csv}

\item{file}{path to output}
}
\description{
Converts an xlsx to bibtex
}
