% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convidence_convert.R
\name{covidence_convert}
\alias{covidence_convert}
\title{covience_convert}
\usage{
covidence_convert(string)
}
\arguments{
\item{string}{String from covidence
fulltext_screened --------}
}
\description{
Covidence convert
}
