% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_difference.R
\name{extract_difference.ris}
\alias{extract_difference.ris}
\title{extract_difference.ris}
\usage{
extract_difference.ris(ris1, ris2, output, threshold = 3)
}
\arguments{
\item{ris1}{path to the ris file to extract references not in the second}

\item{ris2}{path to the reference ris file}

\item{output}{path to the output ris file}

\item{threshold}{scalar, when distances between references are bigger than this, they are retained}
}
\description{
Extracts all references in the first argument, not in the second, then outputs them to the third.
}
