% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{keywords}
\alias{keywords}
\title{keywords}
\usage{
keywords(path = NULL, n = 5, topics = 2, corpus = NULL, ...)
}
\arguments{
\item{path}{path to a PDF or folder containing PDFs}

\item{n}{the number of keywords per topic}

\item{topics}{the number of topics}

\item{corpus}{You can provide a vector of}

\item{...}{additional arguments sent to pdf_words}
}
\description{
get keywords
}
