% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreview_oracle.R
\name{oracle.xlsx_to_ris}
\alias{oracle.xlsx_to_ris}
\title{orcale.xlsx_to_ris}
\usage{
oracle.xlsx_to_ris(csv, file)
}
\arguments{
\item{csv}{path to the csv}

\item{file}{path to output}
}
\description{
Converts an xlsx to bibtex
}
