% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_terms.R
\name{search_terms}
\alias{search_terms}
\title{search_terms}
\usage{
search_terms(..., database = c("scopus", "ovid", "ebscohost", "webofscience"))
}
\arguments{
\item{...}{a series of named vectors. They will be joined within by OR and between by AND}
}
\description{
Constucts search terms
}
\details{
indicate adjacency using word1 w\n word2 e.g. ('search w\2 terms)' will return records where 'search' is within two words of 'terms'. This syntax will be modified for each database.
}
