% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_authors.R
\name{simplify_authors}
\alias{simplify_authors}
\title{simplify_authors}
\usage{
simplify_authors(x)
}
\arguments{
\item{x}{a character vector of comma separated author surnames}
}
\description{
Reduce a list of authors to et al where appropriate
}
