% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma.R
\name{prisma}
\alias{prisma}
\title{prisma}
\usage{
prisma(
  database_records = 0,
  additional_records = 0,
  after_duplicates_removed = 0,
  fulltext_screened = 0,
  final = 0,
  reasons = list(`reason 1` = 0, `reason 2` = 0),
  synthesis_description = "Studies included in\\nquantitative synthesis",
  synthesis_note = "",
  align = "r",
  path = NULL
)
}
\arguments{
\item{database_records}{n from original database}

\item{additional_records}{how many additional references}

\item{after_duplicates_removed}{how many remaining after duplicates removed?}

\item{fulltext_screened}{how many were full text screened}

\item{final}{how many were included in final extract}

\item{reasons}{a named list of reasons list("reason" = 0)}

\item{synthesis_description}{Text to describe synthesis}

\item{align}{'r' (right alignment) or 'l' (left alignment)}

\item{path}{logical. If TRUE, diagram will be saved to image at the desired location. Ext must be pdf, svg, or png}
}
\description{
Create a PRISMA diagram for a meta-analysis
}
